added jdk/test/lib/helpers/ClassFileInstaller
